requirements.txt
Flask
numpy
keras
Pillow
gunicorn
opencv-python